fix(client): apply code-review fixes across amounts, i18n, and correctness
- Decimal-ize transaction/statistics/budget amount handling and align test
  assertions (transaction_model, statistics, budget, report).
- Use theme semantic accent colors instead of hardcoded Colors.green
  (recurring_transaction_card).
- Synchronize slang i18n bundles and regenerated strings across all locales.
- Harden chat/verification/notification async flows, error mapping, and
  response parsing.
- Regenerate freezed/codegen outputs to match upstream model changes.

+++ client/lib/features/home/pages/transaction_detail_page.dart
@@ -222,7 +222,7 @@ class TransactionDetailPage extends ConsumerWidget {
                                     borderRadius: BorderRadius.circular(12),
                                   ),
                                   child: AmountText.large(
-                                    amount: transaction.amount,
+                                    amount: transaction.amount.toDouble(),
                                     type: transaction.type,
                                     currency: transaction.currency ?? 'CNY',
                                   ),
